requirements.txt
tavily-python>=0.3.0
anthropic>=0.18.0
praw>=7.7.0
httpx>=0.27.0
python-dotenv>=1.0.0
